+++ src/Symfony/Component/DependencyInjection/Loader/XmlFileLoader.php
@@ -66,6 +66,7 @@ class XmlFileLoader extends FileLoader
             $this->parseDefinitions($xml, $path, $defaults);
         } finally {
             $this->instanceof = [];
+            $this->registerAliasesForSinglyImplementedInterfaces();
         }
     }