+++ src/Symfony/Component/DependencyInjection/Tests/Loader/FileLoaderTest.php
@@ -240,4 +240,10 @@ class TestFileLoader extends FileLoader
     {
         return false;
     }
+
+    public function registerClasses(Definition $prototype, $namespace, $resource, $exclude = null)
+    {
+        parent::registerClasses($prototype, $namespace, $resource, $exclude);
+        $this->registerAliasesForSinglyImplementedInterfaces();
+    }
 }